// game-service/src/modules/game/game.service.d.ts
import { OnModuleInit, OnModuleDestroy } from '@nestjs/common';
import { Server } from 'socket.io';
import { GameSession, GameRules } from './core';
import { Vector } from './utils';
import { PlayerManager, MapManager, BulletManager } from './managers';
import { ClientProxy } from '@nestjs/microservices';
import { MatchMode, MatchType, ExitStatus, AttackType, MatchMakingData, GameData } from "@transcendence/types";
import { MatchResultService } from '../result/match-result.service';
export declare class GameService implements OnModuleInit, OnModuleDestroy {
    private readonly gameRules;
    private readonly mapManager;
    private readonly playerManager;
    private readonly bulletManager;
    private readonly matchResultService;
    private readonly redis;
    private readonly logger;
    private games;
    private socketToGame;
    private userToGameData;
    private server;
    private lastTime;
    private TIME_STEPS;
    private timeAccumulator;
    private isRunning;
    private nextTickTimeout;
    constructor(gameRules: GameRules, mapManager: MapManager, playerManager: PlayerManager, bulletManager: BulletManager, matchResultService: MatchResultService, redis: ClientProxy);
    onModuleInit(): void;
    onModuleDestroy(): Promise<void>;
    private processNextTick;
    gameLoop(): void;
    handlePlayerDisconnect(socketId: string): ExitStatus;
    removeSession(game: GameSession): Promise<void>;
    removePlayerFromSession(socketId: string): void;
    handleInput(socketId: string, input: Vector, attackType: AttackType, playerIndex?: number): void;
    prepareMatch(gameId: string, players: MatchMakingData[], matchMode: MatchMode, matchType: MatchType): ExitStatus;
    processGameMessage(socketId: string, message: string): void;
    getGameBySocket(socketId: string): GameSession | undefined;
    getGameById(gameId: string): GameSession | undefined;
    setServer(server: Server): void;
    setSocketToGame(socketId: string, gameId: string): void;
    hasPendingMatch(userDbId: string): GameData | undefined;
    removeOldSocket(socketId: string): void;
}
//# sourceMappingURL=game.service.d.ts.map